api/requirements.txt
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3
psycopg2-binary==2.9.10
requests==2.32.3
